Refactor review question screen
